refactor: extract noise-keyframes plugin into plugins/css/
